Organize some files in service folder
